requirements.txt
fastapi
uvicorn
python-multipart
pillow>=9.0.0
pylibdmtx>=0.1.9
qrcode>=7.3.1
python-barcode>=0.13.1
treepoem>=3.17.0
